Separate hooks UI into separate files
