+++ webview-ui/src/components/chat/ChatRow.tsx
@@ -36,6 +36,7 @@ import { CheckpointControls } from "../common/CheckpointControls"
 import CodeAccordian, { cleanPathPrefix } from "../common/CodeAccordian"
 import { ErrorBlockTitle } from "./ErrorBlockTitle"
 import ErrorRow from "./ErrorRow"
+import HookMessage from "./HookMessage"
 import NewTaskPreview from "./NewTaskPreview"
 import QuoteButton from "./QuoteButton"
 import ReportBugPreview from "./ReportBugPreview"
@@ -1637,344 +1638,8 @@ export const ChatRowContent = memo(
 								</div>
 							)
 						}
-					case "hook": {
-						// Parse hook output similar to command output
-						const splitMessage = (text: string) => {
-							const outputIndex = text.indexOf(HOOK_OUTPUT_STRING)
-							if (outputIndex === -1) {
-								return { metadata: text, output: "" }
-							}
-							return {
-								metadata: text.slice(0, outputIndex).trim(),
-								output: text
-									.slice(outputIndex + HOOK_OUTPUT_STRING.length)
-									.trim()
-									.split("")
-									.map((char) => {
-										switch (char) {
-											case "\t":
-												return "→   "
-											case "\b":
-												return "⌫"
-											case "\f":
-												return "⏏"
-											case "\v":
-												return "⇳"
-											default:
-												return char
-										}
-									})
-									.join(""),
-							}
-						}
-
-						const { metadata: metadataStr, output } = splitMessage(message.text || "")
-
-						// Parse the metadata JSON
-						let hookMetadata: {
-							hookName: string
-							toolName?: string
-							status: string
-							exitCode?: number
-							hasJsonResponse?: boolean
-							shouldContinue?: boolean
-							pendingToolInfo?: {
-								tool: string
-								path?: string
-								command?: string
-								content?: string
-								diff?: string
-								regex?: string
-								url?: string
-								mcpTool?: string
-								mcpServer?: string
-								resourceUri?: string
-							}
-						}
-						try {
-							hookMetadata = JSON.parse(metadataStr)
-						} catch {
-							// If parsing fails, still show something
-							hookMetadata = { hookName: "Unknown", status: "unknown" }
-						}
-
-						const isRunning = hookMetadata?.status === "running"
-						const isCompleted = hookMetadata?.status === "completed"
-						const isFailed = hookMetadata?.status === "failed"
-						const isCancelled = hookMetadata?.status === "cancelled"
-						const pendingToolInfo = hookMetadata?.pendingToolInfo
-						const cancelledColor = "var(--vscode-descriptionForeground)"
-
-						return (
-							<>
-								<div style={headerStyle}>
-									<span
-										className="codicon codicon-symbol-event"
-										style={{
-											color: normalColor,
-											marginBottom: "-1.5px",
-										}}></span>
-									<span style={{ color: normalColor, fontWeight: "bold" }}>Hook:</span>
-									<span style={{ color: normalColor }}>{hookMetadata.hookName}</span>
-									{hookMetadata.toolName && (
-										<span style={{ color: "var(--vscode-descriptionForeground)", fontSize: "0.9em" }}>
-											({hookMetadata.toolName})
-										</span>
-									)}
-								</div>
-								<div
-									style={{
-										borderRadius: 6,
-										border: "1px solid var(--vscode-editorGroup-border)",
-										overflow: "hidden",
-										backgroundColor: CHAT_ROW_EXPANDED_BG_COLOR,
-										transition: "all 0.3s ease-in-out",
-									}}>
-									<div
-										style={{
-											display: "flex",
-											alignItems: "center",
-											justifyContent: "space-between",
-											padding: "8px 10px",
-											backgroundColor: CHAT_ROW_EXPANDED_BG_COLOR,
-											borderBottom:
-												pendingToolInfo || output.length > 0
-													? "1px solid var(--vscode-editorGroup-border)"
-													: "none",
-											borderTopLeftRadius: "6px",
-											borderTopRightRadius: "6px",
-										}}>
-										<div
-											style={{
-												display: "flex",
-												alignItems: "center",
-												gap: "8px",
-												flex: 1,
-												minWidth: 0,
-											}}>
-											<div
-												style={{
-													width: "8px",
-													height: "8px",
-													borderRadius: "50%",
-													backgroundColor: isRunning
-														? successColor
-														: isFailed
-															? errorColor
-															: successColor,
-													animation: isRunning ? "pulse 2s ease-in-out infinite" : "none",
-													flexShrink: 0,
-												}}
-											/>
-											<span
-												style={{
-													color: isRunning
-														? successColor
-														: isFailed
-															? errorColor
-															: isCancelled
-																? cancelledColor
-																: successColor,
-													fontWeight: 500,
-													fontSize: "13px",
-													flexShrink: 0,
-												}}>
-												{isRunning
-													? "Running"
-													: isFailed
-														? "Failed"
-														: isCancelled
-															? "Cancelled"
-															: isCompleted
-																? "Completed"
-																: "Unknown"}
-											</span>
-											{hookMetadata.exitCode !== undefined && hookMetadata.exitCode !== 0 && (
-												<span
-													style={{
-														color: "var(--vscode-descriptionForeground)",
-														fontSize: "12px",
-													}}>
-													(exit: {hookMetadata.exitCode})
-												</span>
-											)}
-											{hookMetadata.shouldContinue === false && (
-												<span
-													style={{
-														color: "var(--vscode-descriptionForeground)",
-														fontSize: "12px",
-													}}>
-													(shouldContinue: false)
-												</span>
-											)}
-										</div>
-										{isRunning && (
-											<button
-												onClick={(e) => {
-													e.stopPropagation()
-													TaskServiceClient.cancelHookExecution({}).catch((err) =>
-														console.error("Failed to cancel hook:", err),
-													)
-												}}
-												onMouseEnter={(e) => {
-													e.currentTarget.style.background =
-														"var(--vscode-button-secondaryHoverBackground)"
-												}}
-												onMouseLeave={(e) => {
-													e.currentTarget.style.background = "var(--vscode-button-secondaryBackground)"
-												}}
-												style={{
-													background: "var(--vscode-button-secondaryBackground)",
-													color: "var(--vscode-button-secondaryForeground)",
-													border: "none",
-													borderRadius: "2px",
-													padding: "4px 10px",
-													fontSize: "12px",
-													cursor: "pointer",
-													fontFamily: "inherit",
-												}}>
-												cancel
-											</button>
-										)}
-									</div>
-									{/* Show pending tool info when hook is running */}
-									{isRunning && pendingToolInfo && (
-										<div
-											style={{
-												padding: "12px",
-												backgroundColor: "var(--vscode-editor-background)",
-												borderBottom:
-													output.length > 0 ? "1px solid var(--vscode-editorGroup-border)" : "none",
-												opacity: 0.8,
-											}}>
-											<div style={{ marginBottom: 6 }}>
-												<span style={{ fontWeight: 500 }}>Tool:</span>
-												<span style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-													{pendingToolInfo.tool}
-												</span>
-											</div>
-											{pendingToolInfo.path && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>Path:</span>
-													<span
-														className="ph-no-capture"
-														style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-														{pendingToolInfo.path}
-													</span>
-												</div>
-											)}
-											{pendingToolInfo.command && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>Command:</span>
-													<span
-														className="ph-no-capture"
-														style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-														{pendingToolInfo.command}
-													</span>
-												</div>
-											)}
-											{pendingToolInfo.content && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>Content Preview:</span>
-													<div
-														style={{
-															marginTop: 4,
-															padding: 6,
-															backgroundColor: CODE_BLOCK_BG_COLOR,
-															borderRadius: 3,
-															fontFamily: "monospace",
-															fontSize: "0.85em",
-															whiteSpace: "pre-wrap",
-															wordBreak: "break-word",
-														}}>
-														{pendingToolInfo.content}
-														{pendingToolInfo.content.length >= 200 && "..."}
-													</div>
-												</div>
-											)}
-											{pendingToolInfo.diff && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>Diff Preview:</span>
-													<div
-														style={{
-															marginTop: 4,
-															padding: 6,
-															backgroundColor: CODE_BLOCK_BG_COLOR,
-															borderRadius: 3,
-															fontFamily: "monospace",
-															fontSize: "0.85em",
-															whiteSpace: "pre-wrap",
-															wordBreak: "break-word",
-														}}>
-														{pendingToolInfo.diff}
-														{pendingToolInfo.diff.length >= 200 && "..."}
-													</div>
-												</div>
-											)}
-											{pendingToolInfo.regex && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>Regex:</span>
-													<span
-														className="ph-no-capture"
-														style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-														{pendingToolInfo.regex}
-													</span>
-												</div>
-											)}
-											{pendingToolInfo.url && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>URL:</span>
-													<span
-														className="ph-no-capture"
-														style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-														{pendingToolInfo.url}
-													</span>
-												</div>
-											)}
-											{pendingToolInfo.mcpServer && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>MCP Server:</span>
-													<span
-														className="ph-no-capture"
-														style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-														{pendingToolInfo.mcpServer}
-													</span>
-												</div>
-											)}
-											{pendingToolInfo.mcpTool && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>MCP Tool:</span>
-													<span
-														className="ph-no-capture"
-														style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-														{pendingToolInfo.mcpTool}
-													</span>
-												</div>
-											)}
-											{pendingToolInfo.resourceUri && (
-												<div style={{ marginBottom: 6 }}>
-													<span style={{ fontWeight: 500 }}>Resource URI:</span>
-													<span
-														className="ph-no-capture"
-														style={{ marginLeft: 6, fontFamily: "monospace", fontSize: "0.9em" }}>
-														{pendingToolInfo.resourceUri}
-													</span>
-												</div>
-											)}
-										</div>
-									)}
-									{output.length > 0 && (
-										<CommandOutput
-											isContainerExpanded={true}
-											isOutputFullyExpanded={isHookOutputExpanded}
-											onToggle={() => setIsHookOutputExpanded(!isHookOutputExpanded)}
-											output={output}
-										/>
-									)}
-								</div>
-							</>
-						)
-					}
+					case "hook":
+						return <HookMessage CommandOutput={CommandOutput} message={message} />
 					case "hook_output":
 						// hook_output messages are combined with hook messages, so we don't render them separately
 						return null
